save drawn image + docker
save-endpoint, is a POST-endpoint but as it takes id it will work as a PUT. Had problem using PUT in the frontend -> image-service

also docker build now, however i needed to set host and port to match that of sql in docker. So this wont run locally right now.

To run locally change to 127.0.0.1 in host and 3306 in port. Port can actually be removed when running locally